test: move this PR's adapter tests under the line-limit guardrail
test_adapters.py had grown to 1582 lines — past the first-party 1500-line
guardrail in test_core_decomposition_maintainability — because this PR's
bot rounds kept appending timeout-classification tests to it.

Move exactly those additions into test_adapters_parts/test_adapters_part_001.py,
following the existing *_parts/ convention: the five cancellation/spawn-failure
runner doubles and the sampler stub, plus the twelve tests covering how the
watchdog classification survives cleanup cancellation, cleanup spawn failure,
and the usage-finalize hop. Shared fixtures (_PROMPT, the compose constants,
_CancellingStreamingRunner) are imported from test_adapters, as
test_cursor_gemini_adapters already does.

Test bodies are byte-identical to the originals; no behaviour change. Both
modules now sit under 1050 and 580 lines so later review rounds have headroom.
